refactor: migrate transaction related stuff
